Implementing equals and hashCode
This is to allow you to use Enrollment objects in a java Set and have it eliminate duplicates correctly. I used only fields that would always be present, since some require an admin oAuth token to see and we could end up comparing a copy of the Enrollment we retrieved with and Admin token against one where we didn't use an admin Token and we'd want them to match.
I chose to not use the enrollmentId, because while matching enrollmentId values would definitely indicate a match, non-matching ids (especially with a null value) don't necessarily indicate either way, because we want to be also be able use this to prevent duplicate values where either or both are not yet persisted and so they wouldn't have an enrollment_id. Although I'm not sure that's actually possible considering we get them via Canvas api, but this is my standard way of implementing equals and hashcode so I'm just going to do it this way anyway.
This was needed for Teval, which is already making use of a Set<Enrollment> in its roster update code.
Leankit ID 300