Implementing equals and hashCode
This is to allow you to use User objects in a java Set and have it eliminate duplicates correctly. I used only fields that would always be present, since some require an admin oAuth token to see and we could end up comparing a copy of the User we retrieved with and Admin token against one where we didn't use an admin Token and we'd want them to match.

This was needed for Teval, which is already making use of a Set<Enrollment> in its roster update code. The equals and hashcode of Enrollment relies on the User equals and hashcode implementation.
Leankit ID 300