Refactor BaseImplUTest to not use hardcoded urls
BaseImplUTest was using hardcoded urls when testing methods. This made some of the functionallity (such as readAsUser) difficult to test since the hardcoded url bypasses that functionallity.

I added a public method 'buildTestUrl' which is used in TestCanvasReader and BaseImplUTest to generate the urls. The urls defined in BaseImplUTest were changed to IDs that are eventually passed to the new method.

1 downside of this approach is that it reduces the likelyhood of catching an error in buildCanvasUrl. When BaseImplUTest was originally written there was less functionallity in the class so the old approach was effective. More unit tests around url creation should be added.