Add proof-of-concept for handling Thread-safety for masquerade functionality
This is a small POC to see if this is how we want to handle concurrency issues in BaseImpl. A proxy is introduced to handle masquerade functionality. This is currently very limited and only handles the method in TestCanvasReader.

See the now passing BaseImplUTest#readAsCanvasUserIsNotObviouslyNotThreadsafe() to see that this works