Support options when creating a user.
The creation of a user through the API supports options and this allows more control over how the user is created. As these don’t belong directly on the user object a new method for creating a user has been added and this takes the options as an additional parameter.