Switch the user class ID to be a long.
In canvas IDs for users are normally small and fit into ints but under some circumstances they can become very long (when multiple canvas instances get joined together), in those situations the ID for the instance gets prefixed onto other IDs. So for example a user with an ID of 73 becomes 115370000000000073, and this no longer fits into an int.

11537 is the instance ID for the old Canvas instance.