Set authentication provider when creating a user
Currently, there is a workaround for this which consists of getting the
login and updating the authentication provider on it (2 additional API
calls). So it's not very urgent, but it's definitely a "nice to have".

It accepts both id and type values (in which case, it will find the
first matching provider, see API doc).

Haven't updated the tests as this field is not actually returned in the
User object.