Add support for migrations by URL
This allows for content migrations to have a URL to import rather than a file. It also fixes the option so that we don't use the user ID to perform a import to a group or account.

This does change the API for content migrations, but it makes it clear that people shouldn't continue to use the API as it was.